+++ Pokemon/Assets/Scripts/Character/Stats/TrainedStats.cs
@@ -5,11 +5,11 @@ namespace PokemonGame.Assets.Scripts.Character.Stats
     [Serializable]
     public class TrainedStats
     {
-        public float Health;
-        public float Attack;
-        public float Defense;
-        public float SpAttack;
-        public float SpDefense;
-        public float Speed;
+        public int Health;
+        public int Attack;
+        public int Defense;
+        public int SpAttack;
+        public int SpDefense;
+        public int Speed;
     }
 }
